nearly works

+++ src/shared/GeneralTypes.h
@@ -7,7 +7,7 @@
 #define SERVER_DEFAULT_PORT 59095
 #define CLIENT_DEFAULT_PORT 59096
 #define DEFAULTSEGMENTSIZE 8000
-typedef int Id;
+typedef uint64_t Id;
 typedef uint64_t Hash;
 typedef std::string IpV4Address;
 typedef std::string IpV6Address;
